requirements.txt
requests
python-dotenv
openai
moviepy
gtts
